Renaming molssi_workflow to seamm and workflow to flowchart
