+++ seamm_widgets/__init__.py
@@ -23,16 +23,16 @@ Q_ = ureg.Quantity
 units_class = ureg('1 km').__class__
 
 # Bring up the classes so that they appear to be directly in
-# the molssi_widgets package.
+# the seamm_widgets package.
 
-from molssi_widgets.mousewheel_support import MousewheelSupport  # nopep8
-from molssi_widgets.scrolled_frame import ScrolledFrame  # nopep8
-from molssi_widgets.scrolled_columns import ScrolledColumns  # nopep8
-from molssi_widgets.labeled_widget import LabeledWidget, align_labels  # nopep8
-from molssi_widgets.labeled_entry import LabeledEntry  # nopep8
-from molssi_widgets.labeled_combobox import LabeledCombobox  # nopep8
-from molssi_widgets.unit_entry import UnitEntry  # nopep8
-from molssi_widgets.unit_combobox import UnitCombobox  # nopep8
+from seamm_widgets.mousewheel_support import MousewheelSupport  # nopep8
+from seamm_widgets.scrolled_frame import ScrolledFrame  # nopep8
+from seamm_widgets.scrolled_columns import ScrolledColumns  # nopep8
+from seamm_widgets.labeled_widget import LabeledWidget, align_labels  # nopep8
+from seamm_widgets.labeled_entry import LabeledEntry  # nopep8
+from seamm_widgets.labeled_combobox import LabeledCombobox  # nopep8
+from seamm_widgets.unit_entry import UnitEntry  # nopep8
+from seamm_widgets.unit_combobox import UnitCombobox  # nopep8
 
 
 default_units = {
