+++ seamm_widgets/labeled_combobox.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Labeled entry widget.
+"""Labeled combobox widget.
 
 The goal of these widgets is twofold: to make it easier for developers
 to implement dialogs with compound widgets, and to naturally
@@ -8,7 +8,7 @@ standardize the user interface presented to the user.
 """
 
 import logging
-import molssi_widgets as mw
+import seamm_widgets as sw
 import tkinter as tk
 import tkinter.ttk as ttk
 
@@ -16,46 +16,49 @@ logger = logging.getLogger(__name__)
 
 
 options = {
-    'entry': {
+    'combobox': {
         'class_': 'class_',
         'cursor': 'cursor',
         'exportselection': 'exportselection',
-        'font': 'font',
-        'invalidcommand': 'invalidcommand',
         'justify': 'justify',
-        'show': 'show',
+        'height': 'height',
+        'postcommand': 'postcommand',
+        'state': 'state',
         'style': 'style',
         'takefocus': 'takefocus',
         'variable': 'textvariable',
-        'validate': 'validate',
-        'validatecommand': 'validatecommand',
-        'width': 'width',
-        'xscrollcommand': 'xscrollcommand'
+        'values': 'values',
+        'width': 'width'
     },
 }
 
-class LabeledEntry(mw.LabeledWidget):
+
+class LabeledCombobox(sw.LabeledWidget):
     def __init__(self, parent, *args, **kwargs):
         """Initialize the instance
         """
-        class_ = kwargs.pop('class_', 'MLabeledEntry')
+        class_ = kwargs.pop('class_', 'MLabeledCombobox')
         super().__init__(parent, class_=class_)
 
         interior = self.interior
 
-        # entry
-        justify = kwargs.pop('justify', tk.LEFT)
-        entrywidth = kwargs.pop('width', 15)
+        # combobox
+        height = kwargs.pop('height', 7)
+        width = kwargs.pop('width', 20)
+        state = kwargs.pop('state', 'normal')
 
-        self.entry = ttk.Entry(
+        self.combobox = ttk.Combobox(
             interior,
-            justify=justify,
-            width=entrywidth
+            height=height,
+            width=width,
+            state=state
         )
-        self.entry.grid(row=0, column=0, sticky=tk.EW)
+        self.combobox.grid(row=0, column=0, sticky=tk.EW)
 
         # interior frame
-        self.interior = ttk.Frame(interior)
+        self.interior = ttk.Frame(
+            interior
+        )
         self.interior.grid(row=0, column=1, sticky=tk.NSEW)
 
         interior.columnconfigure(0, weight=1)
@@ -78,37 +81,36 @@ class LabeledEntry(mw.LabeledWidget):
 
         show_all = (len(args) == 0 or args[0] == 'all')
         
-        if show_all or 'entry' in args:
-            self.entry.grid(row=0, column=0, sticky=tk.EW)
+        if show_all or 'combobox' in args:
+            self.combobox.grid(row=0, column=1, sticky=tk.EW)
         else:
-            self.entry.grid_forget()
-            
-    def set(self, value):
-        """Set the value of the entry widget"""
+            self.combobox.grid_forget()
 
-        self.entry.delete(0, tk.END)
+    def set(self, value):
+        """Set the value of the widget"""
+
         if value is None:
             return
 
-        self.entry.insert(0, value)
+        self.combobox.set(value)
 
     def get(self):
         """return the current value"""
-        value = self.entry.get()
+        value = self.combobox.get()
         return value
 
     def config(self, **kwargs):
         """Set the configuration of the megawidget"""
 
         # our options that we deal with
-        entry = options['entry']
+        combobox = options['combobox']
 
         # cannot modify kwargs while iterating over it...
         keys = [*kwargs.keys()]
         for k in keys:
-            if k in entry:
+            if k in combobox:
                 v = kwargs.pop(k)
-                self.entry.config(**{entry[k]: v})
+                self.combobox.config(**{combobox[k]: v})
 
         # having removed our options, pass rest to parent
         super().config(**kwargs)
