+++ seamm_widgets/labeled_entry.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Labeled combobox widget.
+"""Labeled entry widget.
 
 The goal of these widgets is twofold: to make it easier for developers
 to implement dialogs with compound widgets, and to naturally
@@ -8,7 +8,7 @@ standardize the user interface presented to the user.
 """
 
 import logging
-import molssi_widgets as mw
+import seamm_widgets as sw
 import tkinter as tk
 import tkinter.ttk as ttk
 
@@ -16,49 +16,46 @@ logger = logging.getLogger(__name__)
 
 
 options = {
-    'combobox': {
+    'entry': {
         'class_': 'class_',
         'cursor': 'cursor',
         'exportselection': 'exportselection',
+        'font': 'font',
+        'invalidcommand': 'invalidcommand',
         'justify': 'justify',
-        'height': 'height',
-        'postcommand': 'postcommand',
-        'state': 'state',
+        'show': 'show',
         'style': 'style',
         'takefocus': 'takefocus',
         'variable': 'textvariable',
-        'values': 'values',
-        'width': 'width'
+        'validate': 'validate',
+        'validatecommand': 'validatecommand',
+        'width': 'width',
+        'xscrollcommand': 'xscrollcommand'
     },
 }
 
-
-class LabeledCombobox(mw.LabeledWidget):
+class LabeledEntry(sw.LabeledWidget):
     def __init__(self, parent, *args, **kwargs):
         """Initialize the instance
         """
-        class_ = kwargs.pop('class_', 'MLabeledCombobox')
+        class_ = kwargs.pop('class_', 'MLabeledEntry')
         super().__init__(parent, class_=class_)
 
         interior = self.interior
 
-        # combobox
-        height = kwargs.pop('height', 7)
-        width = kwargs.pop('width', 20)
-        state = kwargs.pop('state', 'normal')
+        # entry
+        justify = kwargs.pop('justify', tk.LEFT)
+        entrywidth = kwargs.pop('width', 15)
 
-        self.combobox = ttk.Combobox(
+        self.entry = ttk.Entry(
             interior,
-            height=height,
-            width=width,
-            state=state
+            justify=justify,
+            width=entrywidth
         )
-        self.combobox.grid(row=0, column=0, sticky=tk.EW)
+        self.entry.grid(row=0, column=0, sticky=tk.EW)
 
         # interior frame
-        self.interior = ttk.Frame(
-            interior
-        )
+        self.interior = ttk.Frame(interior)
         self.interior.grid(row=0, column=1, sticky=tk.NSEW)
 
         interior.columnconfigure(0, weight=1)
@@ -81,36 +78,37 @@ class LabeledCombobox(mw.LabeledWidget):
 
         show_all = (len(args) == 0 or args[0] == 'all')
         
-        if show_all or 'combobox' in args:
-            self.combobox.grid(row=0, column=1, sticky=tk.EW)
+        if show_all or 'entry' in args:
+            self.entry.grid(row=0, column=0, sticky=tk.EW)
         else:
-            self.combobox.grid_forget()
-
+            self.entry.grid_forget()
+            
     def set(self, value):
-        """Set the value of the widget"""
+        """Set the value of the entry widget"""
 
+        self.entry.delete(0, tk.END)
         if value is None:
             return
 
-        self.combobox.set(value)
+        self.entry.insert(0, value)
 
     def get(self):
         """return the current value"""
-        value = self.combobox.get()
+        value = self.entry.get()
         return value
 
     def config(self, **kwargs):
         """Set the configuration of the megawidget"""
 
         # our options that we deal with
-        combobox = options['combobox']
+        entry = options['entry']
 
         # cannot modify kwargs while iterating over it...
         keys = [*kwargs.keys()]
         for k in keys:
-            if k in combobox:
+            if k in entry:
                 v = kwargs.pop(k)
-                self.combobox.config(**{combobox[k]: v})
+                self.entry.config(**{entry[k]: v})
 
         # having removed our options, pass rest to parent
         super().config(**kwargs)
