+++ seamm_widgets/mousewheel_support.py
